[ADEN-10273] LogoReplacement template: refactoring
