+++ app/styles/module/_ads.scss
@@ -190,3 +190,11 @@ $closeButtonHeight: 40px;
     }
   }
 }
+
+.new-logo {
+  max-height: 55px;
+}
+
+.tracking-pixel {
+  display: none;
+}
